add itemsCard into itemDisplay
